Restructured tests for router
