+++ internal/testhelpers/testhelpers.go
@@ -5,7 +5,19 @@ import (
 	"time"
 )
 
-const timeout = 5 * time.Second
+// Timeout defines timeout for NoErrorWithinTimeout
+var Timeout = 5 * time.Second
+
+// joinErrChannels multiplexes several error channels
+func joinErrChannels(errChans []<-chan error) chan error {
+	joinedCh := make(chan error)
+	for _, ch := range errChans {
+		go func(errCh <-chan error) {
+			joinedCh <- <-errCh
+		}(ch)
+	}
+	return joinedCh
+}
 
 // NoErrorWithinTimeout tries to read an error from error channel within timeout and returns it.
 // If timeout exceeds, nil value is returned.
@@ -13,7 +25,13 @@ func NoErrorWithinTimeout(ch <-chan error) error {
 	select {
 	case err := <-ch:
 		return err
-	case <-time.After(timeout):
+	case <-time.After(Timeout):
 		return nil
 	}
+}
+
+// NoErrorWithinTimeoutN tries to read an error from error channels within timeout and returns it.
+// If timeout exceeds, nil value is returned.
+func NoErrorWithinTimeoutN(errChans ...<-chan error) error {
+	return NoErrorWithinTimeout(joinErrChannels(errChans))
 }
